Medication, Payment & Care plan details added

+++ app/src/main/java/com/home/asharemedy/view/ActivityAddVitalRecord.kt
@@ -253,7 +253,7 @@ class ActivityAddVitalRecord : BaseActivity() {
                     AppPrefences.getUserID(this), selectedVitalName,
                     Utils.getJSONRequestBodyAny(
                         RequestModel.setVitalRequestModel(
-                            vital_date, vital_reading, vital_unit
+                            vital_date, vital_reading.toFloat(), vital_unit
                         )
                     )
                 )
